+++ lib/image_capture_field.dart
@@ -1,9 +1,11 @@
 library image_capture_field;
 
-import 'dart:io';
+import 'dart:typed_data';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:image_capture_field/crop_the_image/crop_the_image.dart';
 import 'package:image_picker/image_picker.dart';
-import 'controller.dart';
 
 part 'view.dart';
+part 'controller.dart';